data_computation/tp_deployment/model/requirements.txt
# Deep Learning Requirements
torch>=1.7.0
torchvision

# FastAPI Requirements
fastapi
pydantic
python-multipart
uvicorn
requests

# yolo-v5 Requirements
Cython
Pillow
PyYAML
fastapi
matplotlib
numpy
opencv-python-headless
pandas
pillow
seaborn
scipy
tqdm
ipython
psutil

# Your requirements
